requirements.txt
google-api-python-client==1.6.2
google-auth==1.0.2
google-cloud-core==0.28.1
google-cloud-pubsub==0.38.0
mock==2.0.0
oauth2client==3.0.0
requests==2.18.4